Support disk connector

+++ build/multi/arm64-entrypoint.sh
@@ -1,6 +1,7 @@
 #!/bin/sh
 
 run_oss="false"
+run_disk="false"
 
 mkdir -p /var/log/alicloud/
 mkdir -p /host/etc/kubernetes/volumes/disk/uuid
@@ -26,6 +27,7 @@ do
       fi
   elif [ "$item" = "--driver=diskplugin.csi.alibabacloud.com" ]; then
       echo "Running disk plugin...."
+			run_disk="true"
       mkdir -p /var/lib/kubelet/csi-plugins/diskplugin.csi.alibabacloud.com
       rm -rf /var/lib/kubelet/plugins/diskplugin.csi.alibabacloud.com/csi.sock
   elif [ "$item" = "--driver=nasplugin.csi.alibabacloud.com" ]; then
@@ -46,6 +48,7 @@ do
               /usr/bin/nsenter yum install -y fuse-devel
           elif [ "$driver_type" = "disk" ]; then
               echo "Running disk plugin...."
+							run_disk="true"
               mkdir -p /var/lib/kubelet/csi-plugins/diskplugin.csi.alibabacloud.com
               rm -rf /var/lib/kubelet/plugins/diskplugin.csi.alibabacloud.com/csi.sock
           elif [ "$driver_type" = "nas" ]; then
@@ -83,25 +86,26 @@ if [ "$run_oss" = "true" ]; then
             /nsenter --mount=/proc/1/ns/mnt ln -s /usr/bin/ossfs /usr/local/bin/ossfs
         fi
     fi
+fi
 
-    ## install/update csi connector
+if [ "$run_disk" = "true" ] || [ "$run_oss" = "true"]; then
     updateConnector="true"
-    if [ ! -f "/host/etc/csi-tool/csiplugin-connector" ];then
-        mkdir -p /host/etc/csi-tool/
-        echo "mkdir /etc/csi-tool/ directory..."
-    else
-        oldmd5=`md5sum /host/etc/csi-tool/csiplugin-connector | awk '{print $1}'`
-        newmd5=`md5sum /bin/csiplugin-connector | awk '{print $1}'`
-        if [ "$oldmd5" = "$newmd5" ]; then
-            updateConnector="false"
-        else
-            rm -rf /host/etc/csi-tool/
-            rm -rf /host/etc/csi-tool/connector.sock
-            rm -rf /var/log/alicloud/connector.pid
-            mkdir -p /host/etc/csi-tool/
-        fi
-    fi
-
+    if [ ! -f "/host/etc/csi-tool/csiplugin-connector" ]; then
+      mkdir -p /host/etc/csi-tool/
+			echo "mkdir /etc/csi-tool/ directory ...."
+		else
+			oldmd5=`md5sum /host/etc/csi-tool/csiplugin-connector | awk '{print $1}'`
+			newmd5=`md5sum /bin/csiplugin-connector | awk '{print $1}'`
+			if [ "$oldmd5" = "$newmd5" ]; then
+					updateConnector="false"
+			else
+					rm -rf /host/etc/csi-tool/
+					rm -rf /host/etc/csi-tool/connector.sock
+					rm -rf /var/log/alicloud/connector.pid
+					mkdir -p /host/etc/csi-tool/
+			fi
+		fi
+		cp /freezefs.sh /host/etc/csi-tool/freezefs.sh
     if [ "$updateConnector" = "true" ]; then
         echo "Install csiplugin-connector...."
         cp /bin/csiplugin-connector /host/etc/csi-tool/csiplugin-connector
@@ -111,6 +115,12 @@ if [ "$run_oss" = "true" ]; then
 
     # install/update csiplugin connector service
     updateConnectorService="true"
+    if [[ ! -z "${PLUGINS_SOCKETS}" ]];then
+        sed -i 's/Restart=always/Restart=on-failure/g' /bin/csiplugin-connector.service
+        sed -i '/^\[Service\]/a Environment=\"WATCHDOG_SOCKETS_PATH='"${PLUGINS_SOCKETS}"'\"' /bin/csiplugin-connector.service
+        sed -i '/ExecStop=\/bin\/kill -s QUIT $MAINPID/d' /bin/csiplugin-connector.service
+        sed -i '/^\[Service\]/a ExecStop=sh -xc "if [ x$MAINPID != x ]; then /bin/kill -s QUIT $MAINPID; fi"' /bin/csiplugin-connector.service
+    fi
     if [ -f "$systemdDir/csiplugin-connector.service" ];then
         echo "Check csiplugin-connector.service...."
         oldmd5=`md5sum $systemdDir/csiplugin-connector.service | awk '{print $1}'`
@@ -131,8 +141,8 @@ if [ "$run_oss" = "true" ]; then
     rm -rf /var/log/alicloud/connector.pid
     /nsenter --mount=/proc/1/ns/mnt systemctl enable csiplugin-connector.service
     /nsenter --mount=/proc/1/ns/mnt systemctl restart csiplugin-connector.service
-fi
 
+fi
 
 
 # start daemon
